requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
billiard==3.6.4.0
celery==5.2.7
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==4.1.3
djangorestframework==3.14.0
frozenlist==1.3.1
idna==3.4
kombu==5.2.4
multidict==6.0.2
packaging==21.3
prompt-toolkit==3.0.32
pyparsing==3.0.9
pytz==2022.6
redis==4.3.4
six==1.16.0
sqlparse==0.4.3
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
yarl==1.8.1
